apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
motor==3.3.2
pymongo==4.10.1
openai==1.51.0
langchain==0.3.7
langchain-openai==0.2.5
langgraph==0.2.34
langsmith==0.1.129
sentence-transformers==2.7.0
requests==2.32.3
numpy==1.26.4
tiktoken==0.8.0
PyPDF2==3.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
markdown==3.6
aiohttp==3.11.4
httpx==0.27.2
msal==1.28.1
jira==3.8.0
